pyproject.toml
[project]
name = "xworld"
version = "0.2.0"
description = "Cross-domain time series shape clustering — human-in-the-loop exploration"
requires-python = ">=3.13"
dependencies = [
    "streamlit>=1.40.0",
    "anthropic>=0.40.0",
    "ollama>=0.3.0",
    "pandas>=2.2.0",
    "numpy>=2.0.0",
    "scipy>=1.14.0",
    "scikit-learn>=1.5.0",
    "umap-learn>=0.5.7",
    "hdbscan>=0.8.39",
    "matplotlib>=3.9.0",
    "seaborn>=0.13.0",
    "mkdocs>=1.6.0",
    "mkdocs-material>=9.5.0",
    "pyyaml>=6.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
mlflow = ["mlflow>=2.15.0"]

[tool.uv]
dev-dependencies = [
    "ipykernel>=6.29.0",
    "hypothesis>=6.100.0",
    "pytest>=8.0.0",
]
